check if the services are null
the AppContextListener now implements also the ServletRequestListener in order to be able to intercept each request and check if the services are null or not in the session.
If some of them is null, because the container restarted or the application was modified and automatically redeployed, the session is invalidated so that a new session will be created.
Closes #2